输出文件

+++ RegeejairhemFurwhurrahuki/RegeejairhemFurwhurrahuki/Program.cs
@@ -200,16 +200,23 @@ namespace RegeejairhemFurwhurrahuki
             //向标准输入写入要执行的命令。这里使用&是批处理命令的符号，表示前面一个命令不管是否执行成功都执行后面(exit)命令，如果不执行exit命令，后面调用ReadToEnd()方法会假死
             //同类的符号还有&&和||前者表示必须前一个命令执行成功才会执行后面的命令，后者表示必须前一个命令执行失败才会执行后面的命令
 
+            bool exited = false;
+
             // 超时
             Task.Run(() =>
             {
                 Task.Delay(TimeSpan.FromMinutes(1)).ContinueWith(_ =>
                 {
-                    Console.WriteLine("超时");
+                    if (exited)
+                    {
+                        return;
+                    }
+
                     try
                     {
                         if (!p.HasExited)
                         {
+                            Console.WriteLine($"{str} 超时");
                             p.Kill();
                         }
                     }
@@ -236,6 +243,8 @@ namespace RegeejairhemFurwhurrahuki
 
             p.WaitForExit(); //等待程序执行完退出进程
             p.Close();
+
+            exited = true;
 
             return output + "\r\n";
         }
